refactor: move EmoteCategory to platform
