+++ src/platform/item/emote/adr73/emote-data-adr73.ts
@@ -1,4 +1,4 @@
-import { EmoteCategory } from '../../../../dapps/emote-category'
+import { EmoteCategory } from '../emote-category'
 import {
   generateValidator,
   JSONSchema,
