+++ src/platform/item/emote/index.ts
@@ -1,3 +1,4 @@
 export { Emote } from './emote'
 export { EmoteRepresentationADR73 } from './adr73/representation-adr73'
 export { EmoteDataADR73 } from './adr73/emote-data-adr73'
+export { EmoteCategory } from './emote-category'
